fix: use long long instead of int64_t for Integer type

+++ include/groebner_basis/fields/rational.hpp
@@ -8,7 +8,7 @@
 namespace groebner_basis::fields {
 
 class Rational {
-    using Integer = int64_t;
+    using Integer = long long;
 
 public:
     Rational() = default;
